+++ tests/testthat/test-as_pal.R
@@ -1,6 +1,6 @@
-test_that("as_pal_hex works", {
+pal_blues <- palette_bezier(mat_luv_blues)
 
-  pal_blues <- palette_bezier(mat_luv_blues)
+test_that("as_pal_hex works", {
 
   expect_error(
     as_pal_hex("foo"),
@@ -14,6 +14,22 @@ test_that("as_pal_hex works", {
   expect_identical(
     pal_blues_hex(x),
     farver::encode_colour(pal_blues(x), from = "luv")
+  )
+
+})
+
+test_that("as_pal_disc works", {
+
+  n <- 6
+  x <- seq(0, 1, length.out = n)
+
+  pal_blues_disc <- as_pal_disc(pal_blues)
+
+  expect_type(pal_blues_disc, "closure")
+
+  expect_identical(
+    pal_blues(x),
+    pal_blues_disc(n)
   )
 
 })